dict: Tweak to support 4.16+ kernels
In 4.16, rhashtable_walk_start was changed to return void because checking
the return codes for this function didn't bring any value.  This lack of
value is true for older kernels as well, so we'll stop checking now.  That
way we will compile for any kernel.

To read more about the rhashtable_walk_start change, see upstream commit:
97a6ec4ac021f7fbec05c15a3aa0c4aaf0461af5